+++ NEWS.md
@@ -1,8 +1,10 @@
 rd2markdown (dev)
 -----------------
 
+* minor change to S3 method exports (#1 @dgkf)
+
 * minor bug fix to header formatting (additional preceeding newline), which
-  affects a subset of markdown renderers (#1 @dgkf)
+  affects a subset of markdown renderers (#3 @dgkf)
 
 rd2markdown 0.0.1
 -----------------