+++ NEWS.md
@@ -6,6 +6,9 @@ rd2markdown (dev)
 * minor bug fix to header formatting (additional preceeding newline), which
   affects a subset of markdown renderers (#3 @dgkf)
 
+* Add `macros` parameter to `get_rd()` function to allow handling of 
+  custom macros while converting to markdown. (#5 @maksymiuks)
+
 rd2markdown 0.0.1
 -----------------